Change builder class names
