+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/CodeAnalyzer.java
@@ -105,19 +105,19 @@ import io.ballerina.flowmodelgenerator.core.model.ModuleInfo;
 import io.ballerina.flowmodelgenerator.core.model.NodeBuilder;
 import io.ballerina.flowmodelgenerator.core.model.NodeKind;
 import io.ballerina.flowmodelgenerator.core.model.Property;
-import io.ballerina.flowmodelgenerator.core.model.node.Assign;
-import io.ballerina.flowmodelgenerator.core.model.node.BinaryData;
-import io.ballerina.flowmodelgenerator.core.model.node.DataMapper;
-import io.ballerina.flowmodelgenerator.core.model.node.Fail;
-import io.ballerina.flowmodelgenerator.core.model.node.If;
-import io.ballerina.flowmodelgenerator.core.model.node.JsonPayload;
-import io.ballerina.flowmodelgenerator.core.model.node.NewConnection;
-import io.ballerina.flowmodelgenerator.core.model.node.Panic;
-import io.ballerina.flowmodelgenerator.core.model.node.Return;
-import io.ballerina.flowmodelgenerator.core.model.node.Rollback;
-import io.ballerina.flowmodelgenerator.core.model.node.Start;
-import io.ballerina.flowmodelgenerator.core.model.node.Variable;
-import io.ballerina.flowmodelgenerator.core.model.node.XmlPayload;
+import io.ballerina.flowmodelgenerator.core.model.node.AssignBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.BinaryBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.DataMapperBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.FailBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.IfBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.JsonPayloadBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.NewConnectionBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.PanicBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ReturnBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.RollbackBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.StartBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.VariableBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.XmlPayloadBuilder;
 import io.ballerina.flowmodelgenerator.core.utils.ParamUtils;
 import io.ballerina.projects.Project;
 import io.ballerina.tools.text.LinePosition;
@@ -232,8 +232,8 @@ class CodeAnalyzer extends NodeVisitor {
             expr.accept(this);
             if (isNodeUnidentified()) {
                 startNode(NodeKind.RETURN, returnStatementNode)
-                        .metadata().description(String.format(Return.DESCRIPTION, expr)).stepOut()
-                        .properties().expression(expr, Return.RETURN_EXPRESSION_DOC);
+                        .metadata().description(String.format(ReturnBuilder.DESCRIPTION, expr)).stepOut()
+                        .properties().expression(expr, ReturnBuilder.RETURN_EXPRESSION_DOC);
             }
         }
         nodeBuilder.returning();
@@ -769,7 +769,7 @@ class CodeAnalyzer extends NodeVisitor {
     @Override
     public void visit(IfElseStatementNode ifElseStatementNode) {
         startNode(NodeKind.IF, ifElseStatementNode);
-        addConditionalBranch(ifElseStatementNode.condition(), ifElseStatementNode.ifBody(), If.IF_THEN_LABEL);
+        addConditionalBranch(ifElseStatementNode.condition(), ifElseStatementNode.ifBody(), IfBuilder.IF_THEN_LABEL);
         ifElseStatementNode.elseBody().ifPresent(this::analyzeElseBody);
         endNode(ifElseStatementNode);
     }
@@ -786,7 +786,7 @@ class CodeAnalyzer extends NodeVisitor {
             case ELSE_BLOCK -> analyzeElseBody(((ElseBlockNode) elseBody).elseBody());
             case BLOCK_STATEMENT -> {
                 Branch.Builder branchBuilder =
-                        startBranch(If.IF_ELSE_LABEL, NodeKind.ELSE, Branch.BranchKind.BLOCK,
+                        startBranch(IfBuilder.IF_ELSE_LABEL, NodeKind.ELSE, Branch.BranchKind.BLOCK,
                                 Branch.Repeatable.ZERO_OR_ONE);
                 analyzeBlock((BlockStatementNode) elseBody, branchBuilder);
                 endBranch(branchBuilder, elseBody);
@@ -860,12 +860,12 @@ class CodeAnalyzer extends NodeVisitor {
                     .description(description)
                     .stepOut()
                 .codedata()
-                .object(NewConnection.CLIENT_SYMBOL)
-                .symbol(NewConnection.INIT_SYMBOL)
+                .object(NewConnectionBuilder.CLIENT_SYMBOL)
+                .symbol(NewConnectionBuilder.INIT_SYMBOL)
                 .stepOut()
                 .properties()
                 .scope(connectionScope)
-                .checkError(true, NewConnection.CHECK_ERROR_DOC, false);
+                .checkError(true, NewConnectionBuilder.CHECK_ERROR_DOC, false);
         try {
             MethodSymbol methodSymbol =
                     ((ClassSymbol) ((TypeReferenceTypeSymbol) typeSymbol.get()).definition()).initMethod()
@@ -900,7 +900,7 @@ class CodeAnalyzer extends NodeVisitor {
         if (templateExpressionNode.kind() == SyntaxKind.XML_TEMPLATE_EXPRESSION) {
             startNode(NodeKind.XML_PAYLOAD, templateExpressionNode)
                     .metadata()
-                    .description(XmlPayload.DESCRIPTION)
+                    .description(XmlPayloadBuilder.DESCRIPTION)
                     .stepOut()
                     .properties().expression(templateExpressionNode);
         }
@@ -948,9 +948,9 @@ class CodeAnalyzer extends NodeVisitor {
             implicit = true;
             startNode(NodeKind.VARIABLE, variableDeclarationNode)
                     .metadata()
-                    .description(Assign.DESCRIPTION)
+                    .description(AssignBuilder.DESCRIPTION)
                     .stepOut()
-                    .properties().expression((ExpressionNode) null, Variable.EXPRESSION_DOC, true);
+                    .properties().expression((ExpressionNode) null, VariableBuilder.EXPRESSION_DOC, true);
         } else {
             ExpressionNode initializerNode = initializer.get();
             initializerNode.accept(this);
@@ -960,24 +960,24 @@ class CodeAnalyzer extends NodeVisitor {
                 implicit = true;
                 startNode(NodeKind.VARIABLE, variableDeclarationNode)
                         .metadata()
-                        .description(Assign.DESCRIPTION)
+                        .description(AssignBuilder.DESCRIPTION)
                         .stepOut()
-                        .properties().expression(initializerNode, Variable.EXPRESSION_DOC, true);
+                        .properties().expression(initializerNode, VariableBuilder.EXPRESSION_DOC, true);
             }
         }
 
         // TODO: Find a better way on how we can achieve this
-        if (nodeBuilder instanceof DataMapper) {
+        if (nodeBuilder instanceof DataMapperBuilder) {
             nodeBuilder.properties().data(this.typedBindingPatternNode, new HashSet<>());
-        } else if (nodeBuilder instanceof XmlPayload) {
+        } else if (nodeBuilder instanceof XmlPayloadBuilder) {
             nodeBuilder.properties().payload(this.typedBindingPatternNode, "xml");
-        } else if (nodeBuilder instanceof JsonPayload) {
+        } else if (nodeBuilder instanceof JsonPayloadBuilder) {
             nodeBuilder.properties().payload(this.typedBindingPatternNode, "json");
-        } else if (nodeBuilder instanceof BinaryData) {
+        } else if (nodeBuilder instanceof BinaryBuilder) {
             nodeBuilder.properties().payload(this.typedBindingPatternNode, "byte[]");
-        } else if (nodeBuilder instanceof NewConnection) {
-            nodeBuilder.properties().dataVariable(this.typedBindingPatternNode, NewConnection.CONNECTION_NAME_LABEL,
-                    NewConnection.CONNECTION_TYPE_LABEL, false, new HashSet<>());
+        } else if (nodeBuilder instanceof NewConnectionBuilder) {
+            nodeBuilder.properties().dataVariable(this.typedBindingPatternNode, NewConnectionBuilder.CONNECTION_NAME_LABEL,
+                    NewConnectionBuilder.CONNECTION_TYPE_LABEL, false, new HashSet<>());
         } else {
             nodeBuilder.properties().dataVariable(this.typedBindingPatternNode, implicit, new HashSet<>());
         }
@@ -999,15 +999,15 @@ class CodeAnalyzer extends NodeVisitor {
         if (isNodeUnidentified()) {
             startNode(NodeKind.ASSIGN, assignmentStatementNode)
                     .metadata()
-                    .description(Assign.DESCRIPTION)
+                    .description(AssignBuilder.DESCRIPTION)
                     .stepOut()
                     .properties()
-                    .expression(expression, Assign.EXPRESSION_DOC, false)
+                    .expression(expression, AssignBuilder.EXPRESSION_DOC, false)
                     .data(assignmentStatementNode.varRef(), true, new HashSet<>());
         }
 
-        if (nodeBuilder instanceof XmlPayload || nodeBuilder instanceof JsonPayload
-                || nodeBuilder instanceof BinaryData) {
+        if (nodeBuilder instanceof XmlPayloadBuilder || nodeBuilder instanceof JsonPayloadBuilder
+                || nodeBuilder instanceof BinaryBuilder) {
             nodeBuilder.properties().data(assignmentStatementNode.varRef(), false, new HashSet<>());
         }
         endNode(assignmentStatementNode);
@@ -1032,7 +1032,7 @@ class CodeAnalyzer extends NodeVisitor {
     @Override
     public void visit(FailStatementNode failStatementNode) {
         startNode(NodeKind.FAIL, failStatementNode)
-                .properties().expression(failStatementNode.expression(), Fail.FAIL_EXPRESSION_DOC);
+                .properties().expression(failStatementNode.expression(), FailBuilder.FAIL_EXPRESSION_DOC);
         endNode(failStatementNode);
     }
 
@@ -1158,7 +1158,7 @@ class CodeAnalyzer extends NodeVisitor {
     @Override
     public void visit(PanicStatementNode panicStatementNode) {
         startNode(NodeKind.PANIC, panicStatementNode)
-                .properties().expression(panicStatementNode.expression(), Panic.PANIC_EXPRESSION_DOC);
+                .properties().expression(panicStatementNode.expression(), PanicBuilder.PANIC_EXPRESSION_DOC);
         endNode(panicStatementNode);
     }
 
@@ -1171,7 +1171,7 @@ class CodeAnalyzer extends NodeVisitor {
     @Override
     public void visit(StartActionNode startActionNode) {
         startNode(NodeKind.START, startActionNode)
-                .properties().expression(startActionNode.expression(), Start.START_EXPRESSION_DOC);
+                .properties().expression(startActionNode.expression(), StartBuilder.START_EXPRESSION_DOC);
         endNode(startActionNode);
     }
 
@@ -1226,7 +1226,7 @@ class CodeAnalyzer extends NodeVisitor {
         if (optExpr.isPresent()) {
             ExpressionNode expr = optExpr.get();
             expr.accept(this);
-            nodeBuilder.properties().expression(expr, Rollback.ROLLBACK_EXPRESSION_DOC);
+            nodeBuilder.properties().expression(expr, RollbackBuilder.ROLLBACK_EXPRESSION_DOC);
         }
         endNode(rollbackStatementNode);
     }
@@ -1357,7 +1357,7 @@ class CodeAnalyzer extends NodeVisitor {
                 !forceAssign) {
             startNode(NodeKind.JSON_PAYLOAD, constructorExprNode)
                     .metadata()
-                    .description(JsonPayload.DESCRIPTION)
+                    .description(JsonPayloadBuilder.DESCRIPTION)
                     .stepOut()
                     .properties().expression(constructorExprNode);
         }
