+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/ConnectorGenerator.java
@@ -26,7 +26,7 @@ import io.ballerina.flowmodelgenerator.core.model.AvailableNode;
 import io.ballerina.flowmodelgenerator.core.model.Codedata;
 import io.ballerina.flowmodelgenerator.core.model.Metadata;
 import io.ballerina.flowmodelgenerator.core.model.NodeKind;
-import io.ballerina.flowmodelgenerator.core.model.node.NewConnection;
+import io.ballerina.flowmodelgenerator.core.model.node.NewConnectionBuilder;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -75,8 +75,8 @@ public class ConnectorGenerator {
                     .node(NodeKind.NEW_CONNECTION)
                     .org(connectorResult.org())
                     .module(connectorResult.packageName())
-                    .object(NewConnection.CLIENT_SYMBOL)
-                    .symbol(NewConnection.INIT_SYMBOL)
+                    .object(NewConnectionBuilder.CLIENT_SYMBOL)
+                    .symbol(NewConnectionBuilder.INIT_SYMBOL)
                     .id(connectorResult.functionId())
                     .build();
             connectors.add(new AvailableNode(metadata, codedata, true));
