+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/NodeBuilder.java
@@ -24,37 +24,37 @@ import io.ballerina.compiler.api.symbols.ModuleSymbol;
 import io.ballerina.compiler.api.symbols.Symbol;
 import io.ballerina.flowmodelgenerator.core.CommonUtils;
 import io.ballerina.flowmodelgenerator.core.DiagnosticHandler;
-import io.ballerina.flowmodelgenerator.core.model.node.ActionCall;
-import io.ballerina.flowmodelgenerator.core.model.node.Assign;
-import io.ballerina.flowmodelgenerator.core.model.node.BinaryData;
-import io.ballerina.flowmodelgenerator.core.model.node.Break;
-import io.ballerina.flowmodelgenerator.core.model.node.Comment;
-import io.ballerina.flowmodelgenerator.core.model.node.Commit;
-import io.ballerina.flowmodelgenerator.core.model.node.ConfigVariable;
-import io.ballerina.flowmodelgenerator.core.model.node.Continue;
-import io.ballerina.flowmodelgenerator.core.model.node.DataMapper;
-import io.ballerina.flowmodelgenerator.core.model.node.DefaultExpression;
-import io.ballerina.flowmodelgenerator.core.model.node.ErrorHandler;
-import io.ballerina.flowmodelgenerator.core.model.node.EventStart;
-import io.ballerina.flowmodelgenerator.core.model.node.Fail;
-import io.ballerina.flowmodelgenerator.core.model.node.Foreach;
+import io.ballerina.flowmodelgenerator.core.model.node.ActionCallBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.AssignBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.BinaryBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.BreakBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.CommentBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.CommitBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ConfigVariableBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ContinueBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.DataMapperBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ExpressionBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ErrorHandlerBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.EventStartBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.FailBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ForeachBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.FunctionCall;
-import io.ballerina.flowmodelgenerator.core.model.node.If;
-import io.ballerina.flowmodelgenerator.core.model.node.JsonPayload;
-import io.ballerina.flowmodelgenerator.core.model.node.Lock;
-import io.ballerina.flowmodelgenerator.core.model.node.Match;
-import io.ballerina.flowmodelgenerator.core.model.node.NewConnection;
-import io.ballerina.flowmodelgenerator.core.model.node.Panic;
+import io.ballerina.flowmodelgenerator.core.model.node.IfBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.JsonPayloadBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.LockBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.MatchBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.NewConnectionBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.PanicBuilder;
 import io.ballerina.flowmodelgenerator.core.model.node.ResourceActionCall;
-import io.ballerina.flowmodelgenerator.core.model.node.Retry;
-import io.ballerina.flowmodelgenerator.core.model.node.Return;
-import io.ballerina.flowmodelgenerator.core.model.node.Rollback;
-import io.ballerina.flowmodelgenerator.core.model.node.Start;
-import io.ballerina.flowmodelgenerator.core.model.node.Stop;
-import io.ballerina.flowmodelgenerator.core.model.node.Transaction;
-import io.ballerina.flowmodelgenerator.core.model.node.Variable;
-import io.ballerina.flowmodelgenerator.core.model.node.While;
-import io.ballerina.flowmodelgenerator.core.model.node.XmlPayload;
+import io.ballerina.flowmodelgenerator.core.model.node.RetryBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.ReturnBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.RollbackBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.StartBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.StopBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.TransactionBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.VariableBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.WhileBuilder;
+import io.ballerina.flowmodelgenerator.core.model.node.XmlPayloadBuilder;
 import io.ballerina.projects.Document;
 import io.ballerina.tools.text.LinePosition;
 import org.ballerinalang.langserver.commons.workspace.WorkspaceManager;
@@ -92,41 +92,41 @@ public abstract class NodeBuilder implements DiagnosticHandler.DiagnosticCapable
     protected ModuleInfo moduleInfo;
 
     private static final Map<NodeKind, Supplier<? extends NodeBuilder>> CONSTRUCTOR_MAP = new HashMap<>() {{
-        put(NodeKind.IF, If::new);
-        put(NodeKind.RETURN, Return::new);
-        put(NodeKind.EXPRESSION, DefaultExpression::new);
-        put(NodeKind.ERROR_HANDLER, ErrorHandler::new);
-        put(NodeKind.WHILE, While::new);
-        put(NodeKind.CONTINUE, Continue::new);
-        put(NodeKind.BREAK, Break::new);
-        put(NodeKind.PANIC, Panic::new);
-        put(NodeKind.EVENT_START, EventStart::new);
-        put(NodeKind.REMOTE_ACTION_CALL, ActionCall::new);
+        put(NodeKind.IF, IfBuilder::new);
+        put(NodeKind.RETURN, ReturnBuilder::new);
+        put(NodeKind.EXPRESSION, ExpressionBuilder::new);
+        put(NodeKind.ERROR_HANDLER, ErrorHandlerBuilder::new);
+        put(NodeKind.WHILE, WhileBuilder::new);
+        put(NodeKind.CONTINUE, ContinueBuilder::new);
+        put(NodeKind.BREAK, BreakBuilder::new);
+        put(NodeKind.PANIC, PanicBuilder::new);
+        put(NodeKind.EVENT_START, EventStartBuilder::new);
+        put(NodeKind.REMOTE_ACTION_CALL, ActionCallBuilder::new);
         put(NodeKind.RESOURCE_ACTION_CALL, ResourceActionCall::new);
-        put(NodeKind.NEW_CONNECTION, NewConnection::new);
-        put(NodeKind.START, Start::new);
-        put(NodeKind.TRANSACTION, Transaction::new);
-        put(NodeKind.RETRY, Retry::new);
-        put(NodeKind.LOCK, Lock::new);
-        put(NodeKind.FAIL, Fail::new);
-        put(NodeKind.COMMIT, Commit::new);
-        put(NodeKind.ROLLBACK, Rollback::new);
-        put(NodeKind.XML_PAYLOAD, XmlPayload::new);
-        put(NodeKind.JSON_PAYLOAD, JsonPayload::new);
-        put(NodeKind.BINARY_DATA, BinaryData::new);
-        put(NodeKind.STOP, Stop::new);
+        put(NodeKind.NEW_CONNECTION, NewConnectionBuilder::new);
+        put(NodeKind.START, StartBuilder::new);
+        put(NodeKind.TRANSACTION, TransactionBuilder::new);
+        put(NodeKind.RETRY, RetryBuilder::new);
+        put(NodeKind.LOCK, LockBuilder::new);
+        put(NodeKind.FAIL, FailBuilder::new);
+        put(NodeKind.COMMIT, CommitBuilder::new);
+        put(NodeKind.ROLLBACK, RollbackBuilder::new);
+        put(NodeKind.XML_PAYLOAD, XmlPayloadBuilder::new);
+        put(NodeKind.JSON_PAYLOAD, JsonPayloadBuilder::new);
+        put(NodeKind.BINARY_DATA, BinaryBuilder::new);
+        put(NodeKind.STOP, StopBuilder::new);
         put(NodeKind.FUNCTION_CALL, FunctionCall::new);
-        put(NodeKind.FOREACH, Foreach::new);
-        put(NodeKind.DATA_MAPPER, DataMapper::new);
-        put(NodeKind.VARIABLE, Variable::new);
-        put(NodeKind.ASSIGN, Assign::new);
-        put(NodeKind.COMMENT, Comment::new);
-        put(NodeKind.MATCH, Match::new);
-        put(NodeKind.CONFIG_VARIABLE, ConfigVariable::new);
+        put(NodeKind.FOREACH, ForeachBuilder::new);
+        put(NodeKind.DATA_MAPPER, DataMapperBuilder::new);
+        put(NodeKind.VARIABLE, VariableBuilder::new);
+        put(NodeKind.ASSIGN, AssignBuilder::new);
+        put(NodeKind.COMMENT, CommentBuilder::new);
+        put(NodeKind.MATCH, MatchBuilder::new);
+        put(NodeKind.CONFIG_VARIABLE, ConfigVariableBuilder::new);
     }};
 
     public static NodeBuilder getNodeFromKind(NodeKind kind) {
-        return CONSTRUCTOR_MAP.getOrDefault(kind, DefaultExpression::new).get();
+        return CONSTRUCTOR_MAP.getOrDefault(kind, ExpressionBuilder::new).get();
     }
 
     public NodeBuilder setConstData() {
