+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/node/AssignBuilder.java
@@ -19,6 +19,7 @@
 package io.ballerina.flowmodelgenerator.core.model.node;
 
 import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.flowmodelgenerator.core.model.FlowNode;
 import io.ballerina.flowmodelgenerator.core.model.NodeBuilder;
 import io.ballerina.flowmodelgenerator.core.model.NodeKind;
 import io.ballerina.flowmodelgenerator.core.model.Property;
@@ -31,35 +32,49 @@ import java.util.Map;
 import java.util.Optional;
 
 /**
- * Represents the properties of a start node in the flow model.
+ * Represents both variable initialization to a new variable, and assignment of a value to an existing variable.
  *
  * @since 1.4.0
  */
-public class Start extends NodeBuilder {
+public class AssignBuilder extends NodeBuilder {
 
-    public static final String LABEL = "Start";
-    public static final String DESCRIPTION = "Execute a function or a method invocation in a new strand";
-    public static final String START_EXPRESSION_DOC = "Call action or expression";
+    public static final String LABEL = "Assign";
+    public static final String DESCRIPTION = "Assign a value to a variable";
+    public static final String EXPRESSION_DOC = "Assign value";
 
     @Override
     public void setConcreteConstData() {
         metadata().label(LABEL).description(DESCRIPTION);
-        codedata().node(NodeKind.START);
+        codedata().node(NodeKind.ASSIGN);
     }
 
     @Override
     public Map<Path, List<TextEdit>> toSource(SourceBuilder sourceBuilder) {
-        sourceBuilder.token().keyword(SyntaxKind.START_KEYWORD);
-        Optional<Property> property = sourceBuilder.flowNode.getProperty(Property.EXPRESSION_KEY);
-        property.ifPresent(value -> sourceBuilder.token()
+        FlowNode flowNode = sourceBuilder.flowNode;
+
+        Optional<Property> variable = flowNode.getProperty(Property.VARIABLE_KEY);
+        if (variable.isEmpty()) {
+            throw new RuntimeException("Variable is not set for the Assign node");
+        }
+        sourceBuilder.token()
+                .expression(variable.get())
                 .whiteSpace()
-                .expression(value));
-        sourceBuilder.token().endOfStatement();
+                .keyword(SyntaxKind.EQUAL_TOKEN)
+                .whiteSpace();
+
+        Optional<Property> expression = sourceBuilder.flowNode.getProperty(Property.EXPRESSION_KEY);
+        if (expression.isEmpty()) {
+            throw new RuntimeException("Expression is not set for the Assign node");
+        }
+        sourceBuilder.token().expression(expression.get()).endOfStatement();
+
         return sourceBuilder.textEdit(false).build();
     }
 
     @Override
     public void setConcreteTemplateData(TemplateContext context) {
-        properties().expression("", START_EXPRESSION_DOC);
+        properties()
+                .data(null, true, context.getAllVisibleSymbolNames())
+                .expression("", EXPRESSION_DOC);
     }
 }
