+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/node/FailBuilder.java
@@ -18,6 +18,7 @@
 
 package io.ballerina.flowmodelgenerator.core.model.node;
 
+import io.ballerina.compiler.syntax.tree.SyntaxKind;
 import io.ballerina.flowmodelgenerator.core.model.NodeBuilder;
 import io.ballerina.flowmodelgenerator.core.model.NodeKind;
 import io.ballerina.flowmodelgenerator.core.model.Property;
@@ -25,47 +26,40 @@ import io.ballerina.flowmodelgenerator.core.model.SourceBuilder;
 import org.eclipse.lsp4j.TextEdit;
 
 import java.nio.file.Path;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 /**
- * Represents the properties of a comment node.
+ * Represents the properties of a fail node.
  *
  * @since 1.4.0
  */
-public class Comment extends NodeBuilder {
+public class FailBuilder extends NodeBuilder {
 
-    public static final String LABEL = "Comment";
-    public static final String DESCRIPTION = "Comment of a node";
-    private static final String DOUBLE_SLASH = "// ";
-    private static final String NEW_LINE = System.lineSeparator();
+    public static final String LABEL = "Fail";
+    public static final String DESCRIPTION = "Fail the execution";
+    public static final String FAIL_EXPRESSION_DOC = "Fail value";
 
     @Override
     public void setConcreteConstData() {
-        metadata().label(LABEL);
-        codedata().node(NodeKind.COMMENT);
+        metadata().label(LABEL).description(DESCRIPTION);
+        codedata().node(NodeKind.FAIL);
     }
 
     @Override
     public Map<Path, List<TextEdit>> toSource(SourceBuilder sourceBuilder) {
-        Optional<Property> property = sourceBuilder.flowNode.getProperty(Property.COMMENT_KEY);
-        if (property.isEmpty()) {
-            throw new IllegalStateException("Comment must be defined for a comment node");
-        }
-        String formattedComment = Arrays.stream(property.get().toSourceCode().split("\n"))
-                .map(line -> DOUBLE_SLASH + line + NEW_LINE)
-                .collect(Collectors.joining());
-        return sourceBuilder
-                .token().name(NEW_LINE).name(formattedComment).stepOut()
-                .comment()
-                .build();
+        sourceBuilder.token().keyword(SyntaxKind.FAIL_KEYWORD);
+        Optional<Property> property = sourceBuilder.flowNode.getProperty(Property.EXPRESSION_KEY);
+        property.ifPresent(value -> sourceBuilder.token()
+                .whiteSpace()
+                .expression(value));
+        sourceBuilder.token().endOfStatement();
+        return sourceBuilder.textEdit(false).build();
     }
 
     @Override
     public void setConcreteTemplateData(TemplateContext context) {
-        properties().comment("");
+        properties().expression("", FAIL_EXPRESSION_DOC);
     }
 }
