+++ flow-model-generator/modules/flow-model-generator-core/src/main/java/io/ballerina/flowmodelgenerator/core/model/node/MatchBuilder.java
@@ -38,7 +38,7 @@ import java.util.stream.Collectors;
  *
  * @since 1.4.0
  */
-public class Match extends NodeBuilder {
+public class MatchBuilder extends NodeBuilder {
 
     public static final String LABEL = "Match";
     public static final String DESCRIPTION = "Switches the data flow based on the value of an expression.";
